+++ parachain/primitives/beacon/src/config.rs
@@ -1,4 +1,8 @@
 pub const MAX_PROOF_SIZE: u32 = 20;
 
+pub const FEE_RECIPIENT_SIZE: usize = 20;
+pub const EXTRA_DATA_SIZE: usize = 32;
+pub const LOGS_BLOOM_SIZE: usize = 256;
+
 pub const PUBKEY_SIZE: usize = 48;
 pub const SIGNATURE_SIZE: usize = 96;
