+++ parachain/primitives/beacon/src/lib.rs
@@ -1,6 +1,8 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod config;
+pub mod merkleization;
+pub mod ssz;
 
 #[cfg(feature = "std")]
 mod serde_utils;
